+++ src/tests/src/libs/antares/study/parts/hydro/CMakeLists.txt
@@ -1,29 +1,27 @@
 # Useful variables definitions
 set(src_libs_antares_study "${CMAKE_SOURCE_DIR}/libs/antares/study")
 set(src_libs_antares "${CMAKE_SOURCE_DIR}/libs/antares")
+set(src_test_utilis "${CMAKE_SOURCE_DIR}/tests/src/utlis")
 
-# Hydro data transfer
+
+# Hydro data reader
 set(SRC_HYDRO_READER
-	test-hydroreader-class.cpp
-	help-functions.h
-	help-functions.cpp)
+	test-hydroreader-class.cpp)
 
 add_executable(test-hydro-reader ${SRC_HYDRO_READER})
 
 target_include_directories(test-hydro-reader
 	PRIVATE
 	"${src_libs_antares_study}"
-	"${src_libs_antares_study}/parts/hydro"
 	"${src_libs_antares}/array/antares/array"
+	"${src_test_utilis}"
 )
 
 target_link_libraries(test-hydro-reader
 	PRIVATE
 	Boost::unit_test_framework
-	antares-core
-	model_antares
-	antares-solver-hydro
-	yuni-static-core
+	Antares::study
+	test_utils_unit
 )
 
 # Linux
@@ -32,38 +30,30 @@ if(UNIX AND NOT APPLE)
 endif()
 
 set_target_properties(test-hydro-reader PROPERTIES FOLDER Unit-tests/hydro)
-add_test(NAME test-hydro-reader COMMAND test-hydro-reader)
-set_property(TEST test-hydro-reader PROPERTY LABELS unit)
+add_test(NAME hydro-reader COMMAND test-hydro-reader)
+set_property(TEST hydro-reader PROPERTY LABELS unit)
 
-# Useful variables definitions
-set(src_libs_antares_study "${CMAKE_SOURCE_DIR}/libs/antares/study")
-set(src_libs_antares "${CMAKE_SOURCE_DIR}/libs/antares")
 
 # Hydro series
 set(SRC_HYDRO_SERIES
-	test-hydro-series.cpp
-	help-functions.h
-	help-functions.cpp)
+	test-hydro-series.cpp)
 
 add_executable(test-hydro-series ${SRC_HYDRO_SERIES})
 
 target_include_directories(test-hydro-series
 	PRIVATE
 	"${src_libs_antares_study}"
-	"${src_libs_antares_study}/parts/hydro"
-	"${src_libs_antares}/array/antares/array"
 	"${src_libs_antares}/array/antares/array"
 	"${src_libs_antares}/exception/antares/exception"
+	"${src_test_utilis}"
 )
 
 target_link_libraries(test-hydro-series
 	PRIVATE
 	Boost::unit_test_framework
-	antares-core
-	model_antares
-	antares-solver-hydro
-	yuni-static-core
-	exception
+	Antares::study
+	Antares::exception
+	test_utils_unit
 )
 
 # Linux
@@ -72,5 +62,5 @@ if(UNIX AND NOT APPLE)
 endif()
 
 set_target_properties(test-hydro-series PROPERTIES FOLDER Unit-tests/hydro)
-add_test(NAME test-hydro-series COMMAND test-hydro-series)
-set_property(TEST test-hydro-series PROPERTY LABELS unit)
+add_test(NAME hydro-series COMMAND test-hydro-series)
+set_property(TEST hydro-series PROPERTY LABELS unit)
